requirements.txt
langchain>=0.1.0
langgraph>=0.0.10
python-dotenv>=1.0.0
yfinance>=0.2.36
duckduckgo-search>=4.1.1
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
langchain-openai>=0.0.2
langchain-community>=0.0.10 